Updated documentation for Scala body parsers and filters
This does most of #4732, with the exception of
https://www.playframework.com/documentation/2.4.x/HttpApi, which needs
to be fundamentally rewritten.
