+++ documentation/manual/working/scalaGuide/advanced/http/ScalaHttpFilters.md
@@ -15,7 +15,7 @@ In contrast, [[action composition|ScalaActionsComposition]] is intended for rout
 
 ## A simple logging filter
 
-The following is a simple filter that times and logs how long a request takes to execute in Play framework:
+The following is a simple filter that times and logs how long a request takes to execute in Play framework, which implements the [`Filter`](api/scala/play/api/mvc/Filter.html) trait:
 
 @[simple-filter](code/ScalaHttpFilters.scala)
 
@@ -47,12 +47,12 @@ Since filters are applied after routing is done, it is possible to access routin
 
 ## More powerful filters
 
-Play provides a lower level filter API called `EssentialFilter` which gives you full access to the body of the request.  This API allows you to wrap [[EssentialAction|HttpApi]] with another action.
+Play provides a lower level filter API called [`EssentialFilter`](api/scala/play/api/mvc/EssentialFilter.html) which gives you full access to the body of the request.  This API allows you to wrap [[EssentialAction|HttpApi]] with another action.
 
 Here is the above filter example rewritten as an `EssentialFilter`:
 
 @[essential-filter-example](code/EssentialFilter.scala)
 
-The key difference here, apart from creating a new `EssentialAction` to wrap the passed in `next` action, is when we invoke next, we get back an `Iteratee`.  You could wrap this in an `Enumeratee` to do some transformations if you wished.  We then `map` the result of the iteratee and thus handle it.
+The key difference here, apart from creating a new `EssentialAction` to wrap the passed in `next` action, is when we invoke next, we get back an [`Accumulator`](api/scala/play/api/libs/streams/Accumulator.html).  You could compose this with an Akka streams Flow using the `through` method some transformations to the stream if you wished.  We then `map` the result of the iteratee and thus handle it.
 
 > Although it may seem that there are two different filter APIs, there is only one, `EssentialFilter`.  The simpler `Filter` API in the earlier examples extends `EssentialFilter`, and implements it by creating a new `EssentialAction`.  The passed in callback makes it appear to skip the body parsing by creating a promise for the `Result`, while the body parsing and the rest of the action are executed asynchronously.
