+++ documentation/manual/working/scalaGuide/advanced/http/code/ScalaHttpFilters.scala
@@ -7,10 +7,9 @@ import javax.inject.Inject
 import akka.stream.Materializer
 import play.api.Logger
 import play.api.mvc._
-import scala.concurrent.Future
-import play.api.libs.concurrent.Execution.Implicits.defaultContext
+import scala.concurrent.{ExecutionContext, Future}
 
-class LoggingFilter @Inject() (implicit val mat: Materializer) extends Filter {
+class LoggingFilter @Inject() (implicit val mat: Materializer, ec: ExecutionContext) extends Filter {
 
   def apply(nextFilter: RequestHeader => Future[Result])
            (requestHeader: RequestHeader): Future[Result] = {
